Allow to send a list of RSS feeds to Redmine.
It allows me to follow CVEs and create issues directly into redmine from the list of RSS feeds I monitor.
I can then assign the issue to somebody.

Previously I was copying by hand the URL of the CVE and creating the issue by hand too, copying multiple CVE urls into the issue and so on.